+++ sdk/schemaregistry/azure-schemaregistry/CHANGELOG.md
@@ -8,6 +8,10 @@ This version and all future versions will require Python 3.6+. Python 2.7 is no 
 
 - `group_name` and `name` have been added as instance variables to `SchemaProperties`.
 
+### Other Changes
+
+- Updated azure-core minimum dependency to 1.23.0.
+
 ## 1.0.0 (2021-11-10)
 
 **Note:** This is the first stable release of our efforts to create a user-friendly and Pythonic client library for Azure Schema Registry.